refactor: 모임 수요 서비스 로직 분리
