+++ main/src/main/java/org/sopt/makers/crew/main/meetingdemand/v2/service/MeetingDemandV2ServiceImpl.java
@@ -1,22 +1,20 @@
 package org.sopt.makers.crew.main.meetingdemand.v2.service;
 
 import static org.sopt.makers.crew.main.entity.meetingdemand.enums.MeetingDemandStatus.BEFORE_OPEN;
-import static org.sopt.makers.crew.main.global.exception.ErrorStatus.INVALID_MEETING_KEYWORD_SIZE;
 
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+import org.sopt.makers.crew.main.entity.meeting.MeetingRepository;
 import org.sopt.makers.crew.main.entity.meetingdemand.MeetingDemand;
 import org.sopt.makers.crew.main.entity.meetingdemand.MeetingDemandRepository;
 import org.sopt.makers.crew.main.entity.meetingdemand.MeetingDemandWait;
 import org.sopt.makers.crew.main.entity.meetingdemand.MeetingDemandWaitRepository;
-import org.sopt.makers.crew.main.entity.meeting.MeetingRepository;
-import org.sopt.makers.crew.main.entity.tag.enums.MeetingKeywordType;
 import org.sopt.makers.crew.main.entity.user.User;
 import org.sopt.makers.crew.main.entity.user.UserRepository;
-import org.sopt.makers.crew.main.global.exception.BadRequestException;
 import org.sopt.makers.crew.main.global.pagination.dto.PageMetaDto;
+import org.sopt.makers.crew.main.global.pagination.dto.PageOptionsDto;
 import org.sopt.makers.crew.main.meetingdemand.v2.dto.query.MeetingDemandV2GetMeetingDemandsQueryDto;
 import org.sopt.makers.crew.main.meetingdemand.v2.dto.request.MeetingDemandV2CreateMeetingDemandBodyDto;
 import org.sopt.makers.crew.main.meetingdemand.v2.dto.response.MeetingDemandV2CreateMeetingDemandResponseDto;
@@ -36,19 +34,19 @@ import lombok.RequiredArgsConstructor;
 @Transactional(readOnly = true)
 public class MeetingDemandV2ServiceImpl implements MeetingDemandV2Service {
 
-	private static final int MAX_MEETING_KEYWORD_SIZE = 2;
-
 	private final UserRepository userRepository;
 	private final MeetingDemandRepository meetingDemandRepository;
 	private final MeetingDemandWaitRepository meetingDemandWaitRepository;
 	private final MeetingRepository meetingRepository;
+	private final MeetingDemandFactory meetingDemandFactory;
+	private final MeetingDemandPageNormalizer meetingDemandPageNormalizer;
 
 	@Override
 	public MeetingDemandV2GetMeetingDemandsResponseDto getMeetingDemands(
 		MeetingDemandV2GetMeetingDemandsQueryDto queryDto, Integer userId) {
 
 		int totalCount = meetingDemandRepository.countByStatus(BEFORE_OPEN);
-		MeetingDemandV2GetMeetingDemandsQueryDto effectiveQueryDto = adjustPageWhenOutOfRange(queryDto, totalCount);
+		PageOptionsDto effectiveQueryDto = meetingDemandPageNormalizer.normalize(queryDto, totalCount);
 		Page<MeetingDemand> meetingDemands = meetingDemandRepository.findAllByStatus(BEFORE_OPEN,
 			PageRequest.of(
 				effectiveQueryDto.getPage() - 1,
@@ -86,16 +84,7 @@ public class MeetingDemandV2ServiceImpl implements MeetingDemandV2Service {
 	public MeetingDemandV2CreateMeetingDemandResponseDto createMeetingDemand(
 		MeetingDemandV2CreateMeetingDemandBodyDto requestBody, Integer userId) {
 		User user = userRepository.findByIdOrThrow(userId);
-		List<MeetingKeywordType> meetingKeywordTypes = toMeetingKeywordTypes(requestBody.getMeetingKeywordTypes());
-
-		MeetingDemand meetingDemand = MeetingDemand.builder()
-			.user(user)
-			.shortIntro(requestBody.getShortIntro())
-			.expectation(requestBody.getExpectation())
-			.meetingKeywordTypes(meetingKeywordTypes)
-			.joinInfo(requestBody.getJoinInfo())
-			.build();
-
+		MeetingDemand meetingDemand = meetingDemandFactory.create(user, requestBody);
 		MeetingDemand savedMeetingDemand = meetingDemandRepository.save(meetingDemand);
 
 		return MeetingDemandV2CreateMeetingDemandResponseDto.of(savedMeetingDemand.getId());
@@ -152,27 +141,5 @@ public class MeetingDemandV2ServiceImpl implements MeetingDemandV2Service {
 		return meetingDemandWaitRepository.findAllByMeetingDemandIdInAndUserId(meetingDemandIds, userId).stream()
 			.map(MeetingDemandWait::getMeetingDemandId)
 			.collect(Collectors.toSet());
-	}
-
-	private MeetingDemandV2GetMeetingDemandsQueryDto adjustPageWhenOutOfRange(
-		MeetingDemandV2GetMeetingDemandsQueryDto queryDto, int totalCount) {
-		if (totalCount == 0) {
-			return new MeetingDemandV2GetMeetingDemandsQueryDto(1, queryDto.getTake());
-		}
-
-		int pageCount = (int)Math.ceil((double)totalCount / queryDto.getTake());
-		int normalizedPage = Math.min(queryDto.getPage(), pageCount);
-
-		return new MeetingDemandV2GetMeetingDemandsQueryDto(normalizedPage, queryDto.getTake());
-	}
-
-	private List<MeetingKeywordType> toMeetingKeywordTypes(List<String> values) {
-		if (values == null || values.isEmpty() || values.size() > MAX_MEETING_KEYWORD_SIZE) {
-			throw new BadRequestException(INVALID_MEETING_KEYWORD_SIZE.getErrorCode());
-		}
-
-		return values.stream()
-			.map(MeetingKeywordType::ofValue)
-			.toList();
 	}
 }
